requirements.txt
Flask
bootstrap-flask
Flask-WTF
Pillow
requests
beautifulsoup4
